backend/requirements.txt
# ArogyaAI Backend — Production Dependencies
# Core Framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1

# AI Engine
groq==0.9.0

# Database
supabase==2.7.4

# Validation & Schemas
pydantic==2.7.1
pydantic[email]==2.7.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# File uploads
python-multipart==0.0.9

# HTTP Client (async)
httpx==0.27.0

# Rate Limiting
slowapi==0.1.9

# Payment
razorpay==1.4.2

# Redis (optional — for scale)
redis==5.0.6
